+++ src/presentation/controllers/login/login-controller-protocols.ts
@@ -1,4 +1,2 @@
 export * from '../../../domain/usecases/authentication'
 export * from '../../protocols'
-export * from '../../protocols/email-validator'
-export * from '../../protocols/validation'
